+++ samples/generators/jpa/src/main/javagen/topmodel/jpa/sample/demo/daos/utilisateur/UtilisateurDAO.java
@@ -4,11 +4,15 @@
 
 package topmodel.jpa.sample.demo.daos.utilisateur;
 
+import java.util.List;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import topmodel.jpa.sample.demo.dtos.utilisateur.IUtilisateur;
+import topmodel.jpa.sample.demo.entities.utilisateur.TypeUtilisateur.Values;
 import topmodel.jpa.sample.demo.entities.utilisateur.Utilisateur;
-
 
 public interface UtilisateurDAO extends JpaRepository<Utilisateur, Long> {
 
+	List<IUtilisateur> findAllByTypeUtilisateur_Code(Values typeUtilisateurCode);
 }