+++ samples/generators/jpa/src/main/javagen/topmodel/jpa/sample/demo/dtos/securite/ProfilDto.java
@@ -194,9 +194,7 @@ public class ProfilDto implements Serializable {
 	 * @return Une instance de 'Profil'.
 	 */
 	public Profil toProfil(Profil target) {
-		target = target == null ? new Profil() : target;
-		SecuriteMappers.toProfil(this, target);
-		return target;
+		return SecuriteMappers.toProfil(this, target);
 	}
 
 	/**