+++ samples/generators/jpa/src/main/javagen/topmodel/jpa/sample/demo/mappers/securite/securite/SecuriteMapper.java.java
@@ -1,5 +1,0 @@
-////
-//// ATTENTION CE FICHIER EST GENERE AUTOMATIQUEMENT !
-////
-
-package topmodel.jpa.sample.demo.mappers.securite;